[2주차 리팩토링 수정] 결합도 낮추고 응집도 높이도록 수정 문자열 분리와 size검증에 대한 로직은 함께 진행 계산 과정에서 사칙연산과 숫자 검증을 함께 진행하도록 수정
